requirements.txt
absl-py==0.9.0
asgiref==3.2.3
asn1crypto==1.3.0
astor==0.7.1
awscli==1.16.266
backports.tempfile==1.0
backports.weakref==1.0.post1
botocore==1.13.2
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cn2an==0.3.11
colorama==0.4.1
cryptography==2.8
datefinder==0.7.0
dateparser==0.7.2
Django==3.0.2
docutils==0.15.2
et-xmlfile==1.0.1
Flask==1.1.1
future==0.18.2
gast==0.3.3
google-api-core==1.14.3
google-auth==1.9.0
google-cloud-language==1.3.0
google-cloud-speech==1.2.0
googleapis-common-protos==1.51.0
grpcio==1.25.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.3
jmespath==0.9.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
line-bot-sdk==1.15.0
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.4
numpy==1.18.1
openpyxl==3.0.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pydub==0.23.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
regex==2020.1.8
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scipy==1.2.1
six==1.14.0
snownlp==0.12.3
SoundFile==0.10.2
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0

